requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
pydantic>=2.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
PyPDF2>=3.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pymupdf
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
psycopg2-binary
sqlalchemy
authlib
httpx
itsdangerous
PyJWT
starlette
openpyxl
pandas
bcrypt
passlib
python-jose[cryptography]
email-validator
alembic>=1.13.0

# Ground Truth Protocol - Core AI & Database
langchain-openai>=0.1.7
pgvector>=0.2.5
sqlmodel>=0.0.16

# Ground Truth Protocol - Geospatial Intelligence
sentinelhub>=3.10.1
rasterio>=1.3.9
shapely>=2.0.3
torchgeo>=0.5.1
geopy>=2.4.1

# Map visualization & numerical
folium>=0.15.1
numpy>=1.24.0

# Policy Engine & Configuration
pyyaml>=6.0
watchdog>=3.0  # Optional, for policy hot-reload

# LMA Template Generation
python-docx>=1.1.0  # For Word document generation

# Document Retrieval
chromadb>=0.4.0  # Vector database for document similarity search